requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
alembic==1.16.5
anyio==4.12.1
asyncpg==0.31.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpx==0.28.1
httptools==0.7.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
psycopg2-binary==2.9.10
pytest==8.4.2
python-dotenv==1.2.2
python-telegram-bot==22.5
PyYAML==6.0.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
